fix: rename meta-llama to meta_llama
