+++ litellm/llms/meta_llama/chat/transformation.py
@@ -10,7 +10,7 @@ from typing import Optional
 
 from litellm import get_model_info, verbose_logger
 
-from ..openai.chat.gpt_transformation import OpenAIGPTConfig
+from litellm.llms.openai.chat.gpt_transformation import OpenAIGPTConfig
 
 
 class LlamaAPIConfig(OpenAIGPTConfig):
